Initial commit. Part 1 of guide/tutorial complete
This is the initial commit of this repository which
has the first part of the guide fully working.

The guide is a 5 part tutorial on how the create
your own ethernet/ip/tcp stack using tun/tap interface.
Link to the guide: https://www.saminiir.com/lets-code-tcp-ip-stack-1-ethernet-arp/

In this commit where first part is complete the
stack is capable of receiving and replying to
arp (Address Resolution Protocol) ping requests.
More info on about how to run this can be found in
the README.md file.